// src/app/model/request/student-teacher-details-response.ts
export interface StudentTeacherDetailsResponse {
  classId: number;
  classStandard: string;
  classInChargeId: number;
  classInChargeName: string;
  classInChargePhone: string;
  classInChargeEmail: string;
  tamilTeacherId: number;
  tamilTeacherName: string;
  tamilTeacherPhone: string;
  tamilTeacherEmail: string;
  englishTeacherId: number;
  englishTeacherName: string;
  englishTeacherPhone: string;
  englishTeacherEmail: string;
  mathsTeacherId: number;
  mathsTeacherName: string;
  mathsTeacherPhone: string;
  mathsTeacherEmail: string;
  scienceTeacherId: number;
  scienceTeacherName: string;
  scienceTeacherPhone: string;
  scienceTeacherEmail: string;
  socialTeacherId: number;
  socialTeacherName: string;
  socialTeacherPhone: string;
  socialTeacherEmail: string;
}